rename node -> args
