+++ crates/katana/cli/src/lib.rs
@@ -1,10 +1,10 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+pub mod args;
 pub mod file;
-pub mod node;
 pub mod options;
 pub mod utils;
 
+pub use args::NodeArgs;
 pub use file::NodeArgsConfig;
-pub use node::NodeArgs;
 pub use options::*;
